cast

+++ xcsf/env.c
@@ -43,7 +43,7 @@ void env_init(XCSF *xcsf, char **argv)
     char *end;
     if(strcmp(argv[1], "mp") == 0) {
         xcsf->env_vptr = &env_mux_vtbl;
-        env_mux_init(xcsf, strtoimax(argv[2], &end, 10));
+        env_mux_init(xcsf, (int) strtoimax(argv[2], &end, 10));
     } else if(strcmp(argv[1], "maze") == 0) {
         xcsf->env_vptr = &env_maze_vtbl;
         env_maze_init(xcsf, argv[2]);
